pyproject.toml
[project]
name = "mcap-query-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "mcap>=1.1.0",
    "mcap-protobuf-support>=0.5.3",
    "protobuf<5.0.0",
    "psycopg2-binary>=2.9.11",
    "drf-yasg>=1.21.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Dependencies-only layout: backend/frontend/media are not Python packages.
packages = []
